# Precompile one regexp to speedup the Sort function