Before:
```
$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/yansal/natsort
BenchmarkSort1-4   	     300	   4935563 ns/op
PASS
ok  	github.com/yansal/natsort	1.933s
```
After:
```
$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/yansal/natsort
BenchmarkSort1-4   	    3000	    357783 ns/op
PASS
ok  	github.com/yansal/natsort	1.119s
```